requirements.txt
# requirements.txt
# This file lists the Python dependencies for the HCP MCP Server.
# Install them using: pip install -r requirements.txt

fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0